renaming resource into metadata to be more comprehensive
